+++ packages/terra-table/src/index.js
@@ -5,8 +5,9 @@ import columnShape from './proptypes/columnShape';
 import rowShape from './proptypes/rowShape';
 import sectionShape from './proptypes/sectionShape';
 import validateRowHeaderIndex from './proptypes/validators';
+import getFocusableElements from './utils/focusManagement';
 
 export default Table;
 export {
-  GridContext, GridConstants, TableConstants, RowSelectionModes, cellShape, columnShape, rowShape, sectionShape, validateRowHeaderIndex,
+  GridContext, GridConstants, TableConstants, RowSelectionModes, cellShape, columnShape, rowShape, sectionShape, validateRowHeaderIndex, getFocusableElements,
 };